artem_course/requirements.txt
alembic
asyncpg
celery
fastapi[all]
fastapi-cache2[redis]
flower
greenlet

passlib
pillow
python-jose
sqladmin
sqlalchemy
